fix: ajustando detalhes dos códigos e localização dos diretórios
