+++ assignments/00-hello/submissions/Samuel-Abrao/HelloWorld.java
@@ -1,0 +1,5 @@
+public class HelloWorld{
+    public static void main(String[] args){
+        System.out.println("Hello, World!");
+    }
+}
